requirements.txt
cryptography==3.3.2
python-dotenv==1.0.0